add login to medtech using session Fixes #9

+++ radiologistDoctor/models.py
@@ -5,6 +5,7 @@ class findingsTemplate(models.Model):
     id = models.AutoField(primary_key=True)
     template_name = models.CharField(max_length=255)
     template = models.TextField()
+    doctor = models.CharField(max_length=255, null=True)
 
     @classmethod
     def viewTemplate(cls):
@@ -12,10 +13,14 @@ class findingsTemplate(models.Model):
         return templates
     
     @classmethod
-    def deleteTemplateById(cls, template_id):
+    def deleteTemplateById(cls, template_id, user):
         try:
             template_to_delete = cls.objects.get(id=template_id)
-            template_to_delete.delete()
-            return True, "Template deleted successfully."
+            permission_denied = template_to_delete.doctor != user
+            if permission_denied:
+                return False, "You do not have permission to delete"
+            else:
+                template_to_delete.delete()
+                return True, "Template deleted successfully."
         except cls.DoesNotExist:
             return False, "Template with the given ID does not exist."
